### 1. Explain what the PR does

3b1939bbe **chore(proctree): improve /proc status parsing**
8559df83f **chore(proctree): add benchmark for status parsing**


3b1939bbe **chore(proctree): improve /proc status parsing**

```
This short-circuits the status parsing by avoiding fully parsing all
status fields. It also makes ProcStatus a concrete type that contains
only the fields we are interested in, what reduces the memory footprint.

To reduce the size of the function it now uses a map to lookup the field
handler functions.

---

Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf -bench ^Benchmark_newProcStatus$
github.com/aquasecurity/tracee/pkg/utils/proc -benchtime=10000000x

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/utils/proc
cpu: AMD Ryzen 9 7950X 16-Core Processor
Benchmark_newProcStatus-32 10000000  9044 ns/op  7912 B/op  129 allocs/op
PASS
ok      github.com/aquasecurity/tracee/pkg/utils/proc   90.444s

---

| Metric                   | Old Bench. | New Bench. | Improvement (%) |
|--------------------------|------------|------------|-----------------|
| Execution Time (ns/op)   | 13203      | 9044       | ~31.5% faster   |
| Memory Allocations (B/op)| 17851      | 7912       | ~55.7% reduction|
| Allocations (allocs/op)  | 137        | 129        | ~5.8% reduction |
| Total Time  (seconds)    | 132.033    | 90.444     | ~31.5% faster   |
```

8559df83f **chore(proctree): add benchmark for status parsing**

```
Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf -bench ^Benchmark_newProcStatus$
github.com/aquasecurity/tracee/pkg/utils/proc -benchtime=10000000x

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/utils/proc
cpu: AMD Ryzen 9 7950X 16-Core Processor
Benchmark_newProcStatus-32 10000000  13203 ns/op  17851 B/op  137 allocs/op
PASS
ok      github.com/aquasecurity/tracee/pkg/utils/proc   132.033s
```


### 2. Explain how to test it

<!--
Maintainer will review the code, and test the fix/feature, how to run Tracee ?
Give a full command line example and what to look for.
-->

### 3. Other comments

<!--
Links? References? Anything pointing to more context about the change.
-->