@geyslan 2024-10-28
This was tackled by #4364.